feat(v3.1.0): fix auto-grant danh hiệu + 100 Danh Hiệu + MAX_LEVEL 500 tỷ + 2 game (RPS + Nối từ)
Fix CRITICAL bug: hàm met() trong check_and_award dùng match với 25 ID
cố định, mọi danh hiệu mới seed có ID lạ rơi vào _ => false → KHÔNG BAO
GIỜ được auto-grant dù user đạt điều kiện. Fix: mở rộng met() cho 25+100
= 125 ID, mỗi ID có match arm rõ ràng.

Tính năng mới (4):
1. +100 Danh Hiệu (migration 024) — trải đều 16 hạng mục (level/streak/
   comments/games/likes/downloads/followers/reviews/bookmarks/repos/
   news/chat/collections/social/rps/word_chain).
2. MAX_LEVEL 500 Tỷ (500_000_000_000):
   - 12 cấp đầu (xp 0..=12000) dùng LEVELS table (12 tên gọi tĩnh).
   - Cấp 13+ dùng công thức level = 12 + (xp - 12000) / 1000.
   - XP cần cho max: ~5e14 — vừa i64 (BIGINT max 9.2e18).
   - user_xp_totals.total_xp: INT → BIGINT.
3. Game Oẳn tù tì /rps (Kéo búa bao):
   - 3 nút chọn, bot random, +2 XP/win, daily cap 30 ván.
   - 5 huy hiệu RPS (rps_first_win → rps_500_wins).
4. Game Nối từ /word-chain (tiếng Việt):
   - Vocabulary embedded ~370 từ, NFD normalize cho input có dấu.
   - Bot nối tiếp bằng từ bắt đầu bằng chữ cuối của user, +3 XP/valid.
   - 5 huy hiệu word_chain (word_chain_first → word_chain_500).

Refactor (i64 migration): LevelInfo/total_xp/author_xp/PurchaseOutcome
chuyển từ i32 sang i64; level_for/title_for filter parse i64;
collection_limit_for_level nhận i64. 4 test fix cho boundary mới.

Migration 024 mới:
- ALTER TABLE user_xp_totals ALTER COLUMN total_xp TYPE BIGINT.
- INSERT 100 danh hiệu mới (ON CONFLICT DO NOTHING).
- CREATE TABLE rps_plays + word_chain_plays + index bổ trợ.

Verify:
- cargo check --all-targets: pass.
- cargo clippy --all-targets -- -D warnings: 0 warning.
- cargo test --all: 337 test pass.
- cargo fmt --check: clean.
- cargo build --release: pass (binary production-ready).
- Rust 1.98.0 exact pin (rust-toolchain.toml).

Refs: #v3.1.0

+++ src/models/retention.rs
@@ -216,10 +216,11 @@ pub struct ShopItemWithStock {
 }
 
 /// Kết quả mua hàng.
+/// v3.1.0 — total_xp: BIGINT (i64) — hỗ trợ level tới 500 tỷ.
 #[derive(Debug, Clone, Serialize)]
 pub struct PurchaseOutcome {
     pub item_id: String,
-    pub total_xp: i32,
+    pub total_xp: i64,
     /// Mystery box: XP nhận được (0 với vật phẩm khác).
     pub mystery_xp: i32,
 }
